--- v2/requirements.txt
@@ -1,3 +1,0 @@
-pydantic~=2.0
-srsly==2.4.6
-codetiming==1.4.0
