+++ v2_0/requirements.txt
@@ -1,0 +1,3 @@
+pydantic==2.0.3
+srsly==2.4.8
+pytest-benchmark==4.0.0
